requirements.txt
# AI & LLM
google-generativeai>=0.3.0

# Web Scraping
firecrawl-py==1.9.0

# Data Validation & Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0

# UI
streamlit

# Utilities
python-dotenv
